+++ azplugins/__init__.py
@@ -6,6 +6,7 @@
 import analyze
 import bond
 import evaporate
+import flow
 import pair
 import restrain
 import special_pair